+++ features/step_definitions/common_step_definition.js
@@ -1,7 +1,13 @@
 'use strict'
 
+const app = require('../../src/app')
+
 module.exports = function () {
   this.World = require('../support/world.js').World
+
+  this.Given(/^The system is ready$/, {timeout: 10 * 1000}, function (callback) {
+    app.isReady().then(() => callback(), callback)
+  })
 
   this.Given(/^I am a valid user with the uid "([^"]*)"$/, function (uid, callback) {
     this.uid = uid